+++ internal/signatures/generate.go
@@ -22,13 +22,13 @@ import (
 	"bytes"
 	"context"
 	"fmt"
+	"github.com/onflow/flow-cli/internal/util"
 	"github.com/onflow/flow-cli/pkg/flowkit/output"
 
 	"github.com/spf13/cobra"
 
 	"github.com/onflow/flow-cli/internal/command"
 	"github.com/onflow/flow-cli/pkg/flowkit"
-	"github.com/onflow/flow-cli/pkg/flowkit/util"
 )
 
 type flagsGenerate struct {
